fix: pool creation issue, remove some unused codes, and hardcoded explorer urls
