+++ src/abis/gtx/clob/PoolManagerABI.ts
@@ -1,318 +1,3 @@
-const abi = [
-  {
-    type: "constructor",
-    inputs: [
-      {
-        name: "_owner",
-        type: "address",
-        internalType: "address",
-      },
-      {
-        name: "_balanceManager",
-        type: "address",
-        internalType: "address",
-      },
-    ],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "createPool",
-    inputs: [
-      {
-        name: "key",
-        type: "tuple",
-        internalType: "struct PoolKey",
-        components: [
-          {
-            name: "baseCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-          {
-            name: "quoteCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-        ],
-      },
-      {
-        name: "_lotSize",
-        type: "uint256",
-        internalType: "uint256",
-      },
-      {
-        name: "_maxOrderAmount",
-        type: "uint256",
-        internalType: "uint256",
-      },
-    ],
-    outputs: [],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "getPool",
-    inputs: [
-      {
-        name: "key",
-        type: "tuple",
-        internalType: "struct PoolKey",
-        components: [
-          {
-            name: "baseCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-          {
-            name: "quoteCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-        ],
-      },
-    ],
-    outputs: [
-      {
-        name: "",
-        type: "tuple",
-        internalType: "struct IPoolManager.Pool",
-        components: [
-          {
-            name: "maxOrderAmount",
-            type: "uint256",
-            internalType: "uint256",
-          },
-          {
-            name: "lotSize",
-            type: "uint256",
-            internalType: "uint256",
-          },
-          {
-            name: "baseCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-          {
-            name: "quoteCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-          {
-            name: "orderBook",
-            type: "address",
-            internalType: "contract IOrderBook",
-          },
-        ],
-      },
-    ],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "getPoolId",
-    inputs: [
-      {
-        name: "key",
-        type: "tuple",
-        internalType: "struct PoolKey",
-        components: [
-          {
-            name: "baseCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-          {
-            name: "quoteCurrency",
-            type: "address",
-            internalType: "Currency",
-          },
-        ],
-      },
-    ],
-    outputs: [
-      {
-        name: "",
-        type: "bytes32",
-        internalType: "PoolId",
-      },
-    ],
-    stateMutability: "pure",
-  },
-  {
-    type: "function",
-    name: "owner",
-    inputs: [],
-    outputs: [
-      {
-        name: "",
-        type: "address",
-        internalType: "address",
-      },
-    ],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "pools",
-    inputs: [
-      {
-        name: "",
-        type: "bytes32",
-        internalType: "PoolId",
-      },
-    ],
-    outputs: [
-      {
-        name: "maxOrderAmount",
-        type: "uint256",
-        internalType: "uint256",
-      },
-      {
-        name: "lotSize",
-        type: "uint256",
-        internalType: "uint256",
-      },
-      {
-        name: "baseCurrency",
-        type: "address",
-        internalType: "Currency",
-      },
-      {
-        name: "quoteCurrency",
-        type: "address",
-        internalType: "Currency",
-      },
-      {
-        name: "orderBook",
-        type: "address",
-        internalType: "contract IOrderBook",
-      },
-    ],
-    stateMutability: "view",
-  },
-  {
-    type: "function",
-    name: "renounceOwnership",
-    inputs: [],
-    outputs: [],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "setRouter",
-    inputs: [
-      {
-        name: "_router",
-        type: "address",
-        internalType: "address",
-      },
-    ],
-    outputs: [],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "function",
-    name: "transferOwnership",
-    inputs: [
-      {
-        name: "newOwner",
-        type: "address",
-        internalType: "address",
-      },
-    ],
-    outputs: [],
-    stateMutability: "nonpayable",
-  },
-  {
-    type: "event",
-    name: "OwnershipTransferred",
-    inputs: [
-      {
-        name: "previousOwner",
-        type: "address",
-        indexed: true,
-        internalType: "address",
-      },
-      {
-        name: "newOwner",
-        type: "address",
-        indexed: true,
-        internalType: "address",
-      },
-    ],
-    anonymous: false,
-  },
-  {
-    type: "event",
-    name: "PoolCreated",
-    inputs: [
-      {
-        name: "id",
-        type: "bytes32",
-        indexed: true,
-        internalType: "PoolId",
-      },
-      {
-        name: "orderBook",
-        type: "address",
-        indexed: true,
-        internalType: "address",
-      },
-      {
-        name: "baseCurrency",
-        type: "address",
-        indexed: false,
-        internalType: "Currency",
-      },
-      {
-        name: "quoteCurrency",
-        type: "address",
-        indexed: false,
-        internalType: "Currency",
-      },
-      {
-        name: "lotSize",
-        type: "uint256",
-        indexed: false,
-        internalType: "uint256",
-      },
-      {
-        name: "maxOrderAmount",
-        type: "uint256",
-        indexed: false,
-        internalType: "uint256",
-      },
-    ],
-    anonymous: false,
-  },
-  {
-    type: "error",
-    name: "InvalidRouter",
-    inputs: [],
-  },
-  {
-    type: "error",
-    name: "OwnableInvalidOwner",
-    inputs: [
-      {
-        name: "owner",
-        type: "address",
-        internalType: "address",
-      },
-    ],
-  },
-  {
-    type: "error",
-    name: "OwnableUnauthorizedAccount",
-    inputs: [
-      {
-        name: "account",
-        type: "address",
-        internalType: "address",
-      },
-    ],
-  },
-] as const;
+const abi = [{"type":"constructor","inputs":[{"name":"_owner","type":"address","internalType":"address"},{"name":"_balanceManager","type":"address","internalType":"address"}],"stateMutability":"nonpayable"},{"type":"function","name":"addCommonIntermediary","inputs":[{"name":"currency","type":"address","internalType":"Currency"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"allCurrencies","inputs":[{"name":"","type":"uint256","internalType":"uint256"}],"outputs":[{"name":"","type":"address","internalType":"Currency"}],"stateMutability":"view"},{"type":"function","name":"commonIntermediaries","inputs":[{"name":"","type":"uint256","internalType":"uint256"}],"outputs":[{"name":"","type":"address","internalType":"Currency"}],"stateMutability":"view"},{"type":"function","name":"createPool","inputs":[{"name":"_baseCurrency","type":"address","internalType":"Currency"},{"name":"_quoteCurrency","type":"address","internalType":"Currency"},{"name":"_tradingRules","type":"tuple","internalType":"struct IOrderBook.TradingRules","components":[{"name":"minTradeAmount","type":"uint128","internalType":"Quantity"},{"name":"minAmountMovement","type":"uint128","internalType":"Quantity"},{"name":"minPriceMovement","type":"uint128","internalType":"Quantity"},{"name":"minOrderSize","type":"uint128","internalType":"Quantity"},{"name":"slippageTreshold","type":"uint8","internalType":"uint8"}]}],"outputs":[{"name":"","type":"bytes32","internalType":"PoolId"}],"stateMutability":"nonpayable"},{"type":"function","name":"createPoolKey","inputs":[{"name":"currency1","type":"address","internalType":"Currency"},{"name":"currency2","type":"address","internalType":"Currency"}],"outputs":[{"name":"","type":"tuple","internalType":"struct PoolKey","components":[{"name":"baseCurrency","type":"address","internalType":"Currency"},{"name":"quoteCurrency","type":"address","internalType":"Currency"}]}],"stateMutability":"pure"},{"type":"function","name":"getAllCurrencies","inputs":[],"outputs":[{"name":"","type":"address[]","internalType":"Currency[]"}],"stateMutability":"view"},{"type":"function","name":"getCommonIntermediaries","inputs":[],"outputs":[{"name":"","type":"address[]","internalType":"Currency[]"}],"stateMutability":"view"},{"type":"function","name":"getPool","inputs":[{"name":"key","type":"tuple","internalType":"struct PoolKey","components":[{"name":"baseCurrency","type":"address","internalType":"Currency"},{"name":"quoteCurrency","type":"address","internalType":"Currency"}]}],"outputs":[{"name":"","type":"tuple","internalType":"struct IPoolManager.Pool","components":[{"name":"baseCurrency","type":"address","internalType":"Currency"},{"name":"quoteCurrency","type":"address","internalType":"Currency"},{"name":"orderBook","type":"address","internalType":"contract IOrderBook"}]}],"stateMutability":"view"},{"type":"function","name":"getPoolId","inputs":[{"name":"key","type":"tuple","internalType":"struct PoolKey","components":[{"name":"baseCurrency","type":"address","internalType":"Currency"},{"name":"quoteCurrency","type":"address","internalType":"Currency"}]}],"outputs":[{"name":"","type":"bytes32","internalType":"PoolId"}],"stateMutability":"pure"},{"type":"function","name":"getPoolLiquidityScore","inputs":[{"name":"currency1","type":"address","internalType":"Currency"},{"name":"currency2","type":"address","internalType":"Currency"}],"outputs":[{"name":"","type":"uint256","internalType":"uint256"}],"stateMutability":"view"},{"type":"function","name":"isCommonIntermediary","inputs":[{"name":"","type":"address","internalType":"Currency"}],"outputs":[{"name":"","type":"bool","internalType":"bool"}],"stateMutability":"view"},{"type":"function","name":"owner","inputs":[],"outputs":[{"name":"","type":"address","internalType":"address"}],"stateMutability":"view"},{"type":"function","name":"poolExists","inputs":[{"name":"currency1","type":"address","internalType":"Currency"},{"name":"currency2","type":"address","internalType":"Currency"}],"outputs":[{"name":"","type":"bool","internalType":"bool"}],"stateMutability":"view"},{"type":"function","name":"poolLiquidity","inputs":[{"name":"","type":"bytes32","internalType":"PoolId"}],"outputs":[{"name":"","type":"uint256","internalType":"uint256"}],"stateMutability":"view"},{"type":"function","name":"pools","inputs":[{"name":"","type":"bytes32","internalType":"PoolId"}],"outputs":[{"name":"baseCurrency","type":"address","internalType":"Currency"},{"name":"quoteCurrency","type":"address","internalType":"Currency"},{"name":"orderBook","type":"address","internalType":"contract IOrderBook"}],"stateMutability":"view"},{"type":"function","name":"registeredCurrencies","inputs":[{"name":"","type":"address","internalType":"Currency"}],"outputs":[{"name":"","type":"bool","internalType":"bool"}],"stateMutability":"view"},{"type":"function","name":"removeCommonIntermediary","inputs":[{"name":"currency","type":"address","internalType":"Currency"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"renounceOwnership","inputs":[],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"setRouter","inputs":[{"name":"_router","type":"address","internalType":"address"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"transferOwnership","inputs":[{"name":"newOwner","type":"address","internalType":"address"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"updatePoolLiquidity","inputs":[{"name":"key","type":"tuple","internalType":"struct PoolKey","components":[{"name":"baseCurrency","type":"address","internalType":"Currency"},{"name":"quoteCurrency","type":"address","internalType":"Currency"}]},{"name":"liquidityScore","type":"uint256","internalType":"uint256"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"event","name":"CurrencyAdded","inputs":[{"name":"currency","type":"address","indexed":false,"internalType":"Currency"}],"anonymous":false},{"type":"event","name":"IntermediaryAdded","inputs":[{"name":"currency","type":"address","indexed":false,"internalType":"Currency"}],"anonymous":false},{"type":"event","name":"IntermediaryRemoved","inputs":[{"name":"currency","type":"address","indexed":false,"internalType":"Currency"}],"anonymous":false},{"type":"event","name":"OwnershipTransferred","inputs":[{"name":"previousOwner","type":"address","indexed":true,"internalType":"address"},{"name":"newOwner","type":"address","indexed":true,"internalType":"address"}],"anonymous":false},{"type":"event","name":"PoolCreated","inputs":[{"name":"poolId","type":"bytes32","indexed":true,"internalType":"PoolId"},{"name":"orderBook","type":"address","indexed":false,"internalType":"address"},{"name":"baseCurrency","type":"address","indexed":false,"internalType":"Currency"},{"name":"quoteCurrency","type":"address","indexed":false,"internalType":"Currency"}],"anonymous":false},{"type":"event","name":"PoolLiquidityUpdated","inputs":[{"name":"poolId","type":"bytes32","indexed":false,"internalType":"PoolId"},{"name":"newLiquidity","type":"uint256","indexed":false,"internalType":"uint256"}],"anonymous":false},{"type":"error","name":"InvalidRouter","inputs":[]},{"type":"error","name":"OwnableInvalidOwner","inputs":[{"name":"owner","type":"address","internalType":"address"}]},{"type":"error","name":"OwnableUnauthorizedAccount","inputs":[{"name":"account","type":"address","internalType":"address"}]}] as const;
 
 export default abi;
