+++ src/hooks/web3/gtx/clob-dex/balance-manager/write/useWithdraw.ts
@@ -1,80 +1,87 @@
-import { writeContract } from '@wagmi/core';
-import { useCallback, useState } from 'react';
-import { wagmiConfig } from '@/configs/wagmi';
-import { TransactionReceipt } from 'viem';
-import { BALANCE_MANAGER_ADDRESS } from '@/constants/contract-address';
 import BalanceManagerABI from '@/abis/gtx/clob/BalanceManagerABI';
-import { waitForTransaction } from '@wagmi/core';
+import { wagmiConfig } from '@/configs/wagmi';
+import { BALANCE_MANAGER_ADDRESS } from '@/constants/contract-address';
 import { HexAddress } from '@/types/general/address';
-import { BaseOptions } from '../read/useBalanceManager';
-import { BalanceOperationParams } from './useDeposit';
+import { waitForTransaction, writeContract } from '@wagmi/core';
+import { useCallback, useState } from 'react';
+import { TransactionReceipt } from 'viem';
+import { useChainId } from 'wagmi';
 
 // Withdraw hook
-interface WithdrawParams extends BalanceOperationParams {
-    user?: HexAddress;
-  }
-  
-  interface UseWithdrawReturn {
-    withdraw: (params: WithdrawParams) => Promise<TransactionReceipt>;
-    isWithdrawing: boolean;
-    error: Error | null;
-  }
-  
-  export const useWithdraw = (options: BaseOptions = {}): UseWithdrawReturn => {
-    const { onSuccess, onError } = options;
-    const [isWithdrawing, setIsWithdrawing] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
-  
-    const withdraw = useCallback(
-      async ({ currency, amount, user }: WithdrawParams): Promise<TransactionReceipt> => {
-        setIsWithdrawing(true);
-        setError(null);
-  
-        try {
-          if (user) {
-            const hash = await writeContract(wagmiConfig, {
-              address: BALANCE_MANAGER_ADDRESS,
-              abi: BalanceManagerABI,
-              functionName: 'withdraw',
-              args: [currency, amount, user] as const,
-            });
-  
-            const receipt = await waitForTransaction(wagmiConfig, {
-              hash,
-            });
-  
-            onSuccess?.(receipt);
-            return receipt;
-          } else {
-            const hash = await writeContract(wagmiConfig, {
-              address: BALANCE_MANAGER_ADDRESS,
-              abi: BalanceManagerABI,
-              functionName: 'withdraw',
-              args: [currency, amount] as const,
-            });
-  
-            const receipt = await waitForTransaction(wagmiConfig, {
-              hash,
-            });
-  
-            onSuccess?.(receipt);
-            return receipt;
-          }
-        } catch (err: unknown) {
-          const error = err instanceof Error ? err : new Error('Failed to withdraw');
-          setError(error);
-          onError?.(error);
-          throw error;
-        } finally {
-          setIsWithdrawing(false);
+interface WithdrawParams {
+  user?: HexAddress;
+  currency: HexAddress;
+  amount: bigint;
+}
+
+interface UseWithdrawReturn {
+  withdraw: (params: WithdrawParams) => Promise<TransactionReceipt>;
+  isWithdrawing: boolean;
+  error: Error | null;
+}
+
+export interface BaseOptions {
+  onSuccess?: (receipt: TransactionReceipt) => void;
+  onError?: (error: Error) => void;
+}
+
+export const useWithdraw = (options: BaseOptions = {}): UseWithdrawReturn => {
+  const { onSuccess, onError } = options;
+  const [isWithdrawing, setIsWithdrawing] = useState(false);
+  const [error, setError] = useState<Error | null>(null);
+
+  const chainId = useChainId()
+
+  const withdraw = useCallback(
+    async ({ currency, amount, user }: WithdrawParams): Promise<TransactionReceipt> => {
+      setIsWithdrawing(true);
+      setError(null);
+
+      try {
+        if (user) {
+          const hash = await writeContract(wagmiConfig, {
+            address: BALANCE_MANAGER_ADDRESS(chainId) as `0x${string}`,
+            abi: BalanceManagerABI,
+            functionName: 'withdraw',
+            args: [currency, amount, user] as const,
+          });
+
+          const receipt = await waitForTransaction(wagmiConfig, {
+            hash,
+          });
+
+          onSuccess?.(receipt);
+          return receipt;
+        } else {
+          const hash = await writeContract(wagmiConfig, {
+            address: BALANCE_MANAGER_ADDRESS(chainId) as `0x${string}`,
+            abi: BalanceManagerABI,
+            functionName: 'withdraw',
+            args: [currency, amount] as const,
+          });
+
+          const receipt = await waitForTransaction(wagmiConfig, {
+            hash,
+          });
+
+          onSuccess?.(receipt);
+          return receipt;
         }
-      },
-      [onSuccess, onError]
-    );
-  
-    return {
-      withdraw,
-      isWithdrawing,
-      error,
-    };
+      } catch (err: unknown) {
+        const error = err instanceof Error ? err : new Error('Failed to withdraw');
+        setError(error);
+        onError?.(error);
+        throw error;
+      } finally {
+        setIsWithdrawing(false);
+      }
+    },
+    [onSuccess, onError]
+  );
+
+  return {
+    withdraw,
+    isWithdrawing,
+    error,
   };
+};
