+++ src/hooks/web3/gtx/perp-dex/market/useMarkets.ts
@@ -5,6 +5,7 @@ import { TransactionReceipt } from 'viem';
 import { POOL_MANAGER_ADDRESS } from '@/constants/contract-address';
 import PoolManagerABI from '@/abis/gtx/clob/PoolManagerABI';
 import { HexAddress } from '@/types/general/address';
+import { useChainId } from 'wagmi';
 
 // Common types
 interface BaseOptions {
@@ -27,399 +28,308 @@ interface Pool {
     orderBook: HexAddress;
 }
 
-// Hook for retrieving contract owner
-interface UsePoolManagerOwnerReturn {
-    getOwner: () => Promise<HexAddress>;
-    isLoading: boolean;
-    error: Error | null;
-}
+// // Hook for creating a new pool
+// interface CreatePoolParams {
+//     poolKey: PoolKey;
+//     lotSize: bigint;
+//     maxOrderAmount: bigint;
+// }
 
-export const usePoolManagerOwner = (): UsePoolManagerOwnerReturn => {
-    const [isLoading, setIsLoading] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
+// interface UseCreatePoolReturn {
+//     createPool: (params: CreatePoolParams) => Promise<TransactionReceipt>;
+//     isCreating: boolean;
+//     error: Error | null;
+// }
 
-    const getOwner = useCallback(async (): Promise<HexAddress> => {
-        setIsLoading(true);
-        setError(null);
+// export const useCreatePool = (options: BaseOptions = {}): UseCreatePoolReturn => {
+//     const { onSuccess, onError } = options;
+//     const [isCreating, setIsCreating] = useState(false);
+//     const [error, setError] = useState<Error | null>(null);
 
-        try {
-            const owner = await readContract(wagmiConfig, {
-                address: POOL_MANAGER_ADDRESS,
-                abi: PoolManagerABI,
-                functionName: 'owner',
-            });
+//     const chainId = useChainId();
 
-            return owner;
-        } catch (err: unknown) {
-            const error = err instanceof Error ? err : new Error('Failed to get pool manager owner');
-            setError(error);
-            throw error;
-        } finally {
-            setIsLoading(false);
-        }
-    }, []);
+//     const createPool = useCallback(
+//         async ({ poolKey, lotSize, maxOrderAmount }: CreatePoolParams): Promise<TransactionReceipt> => {
+//             setIsCreating(true);
+//             setError(null);
 
-    return {
-        getOwner,
-        isLoading,
-        error,
-    };
-};
+//             try {
+//                 const hash = await writeContract(wagmiConfig, {
+//                     address: POOL_MANAGER_ADDRESS(chainId) as `0x${string}`,
+//                     abi: PoolManagerABI,
+//                     functionName: 'createPool',
+//                     args: [poolKey.baseCurrency, lotSize, maxOrderAmount] as const,
+//                 });
 
-// Hook for setting router address
-interface SetRouterParams {
-    router: HexAddress;
-}
+//                 const receipt = await waitForTransaction(wagmiConfig, {
+//                     hash,
+//                 });
 
-interface UseSetRouterReturn {
-    setRouter: (params: SetRouterParams) => Promise<TransactionReceipt>;
-    isSettingRouter: boolean;
-    error: Error | null;
-}
+//                 onSuccess?.(receipt);
+//                 return receipt;
+//             } catch (err: unknown) {
+//                 const error = err instanceof Error ? err : new Error('Failed to create pool');
+//                 setError(error);
+//                 onError?.(error);
+//                 throw error;
+//             } finally {
+//                 setIsCreating(false);
+//             }
+//         },
+//         [onSuccess, onError]
+//     );
 
-export const useSetRouter = (options: BaseOptions = {}): UseSetRouterReturn => {
-    const { onSuccess, onError } = options;
-    const [isSettingRouter, setIsSettingRouter] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
+//     return {
+//         createPool,
+//         isCreating,
+//         error,
+//     };
+// };
 
-    const setRouter = useCallback(
-        async ({ router }: SetRouterParams): Promise<TransactionReceipt> => {
-            setIsSettingRouter(true);
-            setError(null);
+// // Hook for getting pool details
+// interface GetPoolParams {
+//     poolKey: PoolKey;
+// }
 
-            try {
-                const hash = await writeContract(wagmiConfig, {
-                    address: POOL_MANAGER_ADDRESS,
-                    abi: PoolManagerABI,
-                    functionName: 'setRouter',
-                    args: [router] as const,
-                });
+// interface UseGetPoolReturn {
+//     getPool: (params: GetPoolParams) => Promise<Pool>;
+//     isLoading: boolean;
+//     error: Error | null;
+// }
 
-                const receipt = await waitForTransaction(wagmiConfig, {
-                    hash,
-                });
+// export const useGetPool = (): UseGetPoolReturn => {
+//     const [isLoading, setIsLoading] = useState(false);
+//     const [error, setError] = useState<Error | null>(null);
 
-                onSuccess?.(receipt);
-                return receipt;
-            } catch (err: unknown) {
-                const error = err instanceof Error ? err : new Error('Failed to set router');
-                setError(error);
-                onError?.(error);
-                throw error;
-            } finally {
-                setIsSettingRouter(false);
-            }
-        },
-        [onSuccess, onError]
-    );
+//         const chainId = useChainId();
 
-    return {
-        setRouter,
-        isSettingRouter,
-        error,
-    };
-};
 
-// Hook for creating a new pool
-interface CreatePoolParams {
-    poolKey: PoolKey;
-    lotSize: bigint;
-    maxOrderAmount: bigint;
-}
+//     const getPool = useCallback(async ({ poolKey }: GetPoolParams): Promise<Pool> => {
+//         setIsLoading(true);
+//         setError(null);
 
-interface UseCreatePoolReturn {
-    createPool: (params: CreatePoolParams) => Promise<TransactionReceipt>;
-    isCreating: boolean;
-    error: Error | null;
-}
+//         try {
+//             const pool = await readContract(wagmiConfig, {
+//                 address: POOL_MANAGER_ADDRESS(chainId) as `0x${string}`,
+//                 abi: PoolManagerABI,
+//                 functionName: 'getPool',
+//                 args: [{
+//                     baseCurrency: poolKey.baseCurrency,
+//                     quoteCurrency: poolKey.quoteCurrency
+//                 }]
+//             });
 
-export const useCreatePool = (options: BaseOptions = {}): UseCreatePoolReturn => {
-    const { onSuccess, onError } = options;
-    const [isCreating, setIsCreating] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
+//             return pool;
+//         } catch (err: unknown) {
+//             const error = err instanceof Error ? err : new Error('Failed to get pool details');
+//             setError(error);
+//             throw error;
+//         } finally {
+//             setIsLoading(false);
+//         }
+//     }, []);
 
-    const createPool = useCallback(
-        async ({ poolKey, lotSize, maxOrderAmount }: CreatePoolParams): Promise<TransactionReceipt> => {
-            setIsCreating(true);
-            setError(null);
+//     return {
+//         getPool,
+//         isLoading,
+//         error,
+//     };
+// // };
 
-            try {
-                const hash = await writeContract(wagmiConfig, {
-                    address: POOL_MANAGER_ADDRESS,
-                    abi: PoolManagerABI,
-                    functionName: 'createPool',
-                    args: [{
-                        baseCurrency: poolKey.baseCurrency,
-                        quoteCurrency: poolKey.quoteCurrency
-                    }, lotSize, maxOrderAmount] as const,
-                });
+// // Hook for getting pool ID
+// interface GetPoolIdParams {
+//     poolKey: PoolKey;
+// }
 
-                const receipt = await waitForTransaction(wagmiConfig, {
-                    hash,
-                });
+// interface UseGetPoolIdReturn {
+//     getPoolId: (params: GetPoolIdParams) => Promise<string>;
+//     isLoading: boolean;
+//     error: Error | null;
+// }
 
-                onSuccess?.(receipt);
-                return receipt;
-            } catch (err: unknown) {
-                const error = err instanceof Error ? err : new Error('Failed to create pool');
-                setError(error);
-                onError?.(error);
-                throw error;
-            } finally {
-                setIsCreating(false);
-            }
-        },
-        [onSuccess, onError]
-    );
+// export const useGetPoolId = (): UseGetPoolIdReturn => {
+//     const [isLoading, setIsLoading] = useState(false);
+//     const [error, setError] = useState<Error | null>(null);
 
-    return {
-        createPool,
-        isCreating,
-        error,
-    };
-};
+//     const getPoolId = useCallback(async ({ poolKey }: GetPoolIdParams): Promise<string> => {
+//         setIsLoading(true);
+//         setError(null);
 
-// Hook for getting pool details
-interface GetPoolParams {
-    poolKey: PoolKey;
-}
+//         try {
+//             const poolId = await readContract(wagmiConfig, {
+//                 address: POOL_MANAGER_ADDRESS,
+//                 abi: PoolManagerABI,
+//                 functionName: 'getPoolId',
+//                 args: [{
+//                     baseCurrency: poolKey.baseCurrency,
+//                     quoteCurrency: poolKey.quoteCurrency
+//                 }] as const,
+//             });
 
-interface UseGetPoolReturn {
-    getPool: (params: GetPoolParams) => Promise<Pool>;
-    isLoading: boolean;
-    error: Error | null;
-}
+//             return poolId;
+//         } catch (err: unknown) {
+//             const error = err instanceof Error ? err : new Error('Failed to get pool ID');
+//             setError(error);
+//             throw error;
+//         } finally {
+//             setIsLoading(false);
+//         }
+//     }, []);
 
-export const useGetPool = (): UseGetPoolReturn => {
-    const [isLoading, setIsLoading] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
+//     return {
+//         getPoolId,
+//         isLoading,
+//         error,
+//     };
+// };
 
-    const getPool = useCallback(async ({ poolKey }: GetPoolParams): Promise<Pool> => {
-        setIsLoading(true);
-        setError(null);
+// // Hook for getting pool details by ID
+// interface GetPoolByIdParams {
+//     poolId: HexAddress; // bytes32
+// }
 
-        try {
-            const pool = await readContract(wagmiConfig, {
-                address: POOL_MANAGER_ADDRESS,
-                abi: PoolManagerABI,
-                functionName: 'getPool',
-                args: [{
-                    baseCurrency: poolKey.baseCurrency,
-                    quoteCurrency: poolKey.quoteCurrency
-                }]
-            });
+// interface UseGetPoolByIdReturn {
+//     getPoolById: (params: GetPoolByIdParams) => Promise<Pool>;
+//     isLoading: boolean;
+//     error: Error | null;
+// }
 
-            return pool;
-        } catch (err: unknown) {
-            const error = err instanceof Error ? err : new Error('Failed to get pool details');
-            setError(error);
-            throw error;
-        } finally {
-            setIsLoading(false);
-        }
-    }, []);
+// export const useGetPoolById = (): UseGetPoolByIdReturn => {
+//     const [isLoading, setIsLoading] = useState(false);
+//     const [error, setError] = useState<Error | null>(null);
 
-    return {
-        getPool,
-        isLoading,
-        error,
-    };
-};
+//     const getPoolById = useCallback(async ({ poolId }: GetPoolByIdParams): Promise<Pool> => {
+//         setIsLoading(true);
+//         setError(null);
 
-// Hook for getting pool ID
-interface GetPoolIdParams {
-    poolKey: PoolKey;
-}
+//         try {
+//             const poolData = await readContract(wagmiConfig, {
+//                 address: POOL_MANAGER_ADDRESS,
+//                 abi: PoolManagerABI,
+//                 functionName: 'pools',
+//                 args: [poolId] as const,
+//             });
 
-interface UseGetPoolIdReturn {
-    getPoolId: (params: GetPoolIdParams) => Promise<string>;
-    isLoading: boolean;
-    error: Error | null;
-}
+//             // The returned data is a tuple array, convert it to our Pool type
+//             return {
+//                 maxOrderAmount: poolData[0],
+//                 lotSize: poolData[1],
+//                 baseCurrency: poolData[2],
+//                 quoteCurrency: poolData[3],
+//                 orderBook: poolData[4],
+//             };
+//         } catch (err: unknown) {
+//             const error = err instanceof Error ? err : new Error('Failed to get pool by ID');
+//             setError(error);
+//             throw error;
+//         } finally {
+//             setIsLoading(false);
+//         }
+//     }, []);
 
-export const useGetPoolId = (): UseGetPoolIdReturn => {
-    const [isLoading, setIsLoading] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
+//     return {
+//         getPoolById,
+//         isLoading,
+//         error,
+//     };
+// };
 
-    const getPoolId = useCallback(async ({ poolKey }: GetPoolIdParams): Promise<string> => {
-        setIsLoading(true);
-        setError(null);
+// // Hook for transferring ownership
+// interface TransferOwnershipParams {
+//     newOwner: HexAddress;
+// }
 
-        try {
-            const poolId = await readContract(wagmiConfig, {
-                address: POOL_MANAGER_ADDRESS,
-                abi: PoolManagerABI,
-                functionName: 'getPoolId',
-                args: [{
-                    baseCurrency: poolKey.baseCurrency,
-                    quoteCurrency: poolKey.quoteCurrency
-                }] as const,
-            });
+// interface UseTransferOwnershipReturn {
+//     transferOwnership: (params: TransferOwnershipParams) => Promise<TransactionReceipt>;
+//     isTransferring: boolean;
+//     error: Error | null;
+// }
 
-            return poolId;
-        } catch (err: unknown) {
-            const error = err instanceof Error ? err : new Error('Failed to get pool ID');
-            setError(error);
-            throw error;
-        } finally {
-            setIsLoading(false);
-        }
-    }, []);
+// export const useTransferOwnership = (options: BaseOptions = {}): UseTransferOwnershipReturn => {
+//     const { onSuccess, onError } = options;
+//     const [isTransferring, setIsTransferring] = useState(false);
+//     const [error, setError] = useState<Error | null>(null);
 
-    return {
-        getPoolId,
-        isLoading,
-        error,
-    };
-};
+//     const transferOwnership = useCallback(
+//         async ({ newOwner }: TransferOwnershipParams): Promise<TransactionReceipt> => {
+//             setIsTransferring(true);
+//             setError(null);
 
-// Hook for getting pool details by ID
-interface GetPoolByIdParams {
-    poolId: HexAddress; // bytes32
-}
+//             try {
+//                 const hash = await writeContract(wagmiConfig, {
+//                     address: POOL_MANAGER_ADDRESS,
+//                     abi: PoolManagerABI,
+//                     functionName: 'transferOwnership',
+//                     args: [newOwner] as const,
+//                 });
 
-interface UseGetPoolByIdReturn {
-    getPoolById: (params: GetPoolByIdParams) => Promise<Pool>;
-    isLoading: boolean;
-    error: Error | null;
-}
+//                 const receipt = await waitForTransaction(wagmiConfig, {
+//                     hash,
+//                 });
 
-export const useGetPoolById = (): UseGetPoolByIdReturn => {
-    const [isLoading, setIsLoading] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
+//                 onSuccess?.(receipt);
+//                 return receipt;
+//             } catch (err: unknown) {
+//                 const error = err instanceof Error ? err : new Error('Failed to transfer ownership');
+//                 setError(error);
+//                 onError?.(error);
+//                 throw error;
+//             } finally {
+//                 setIsTransferring(false);
+//             }
+//         },
+//         [onSuccess, onError]
+//     );
 
-    const getPoolById = useCallback(async ({ poolId }: GetPoolByIdParams): Promise<Pool> => {
-        setIsLoading(true);
-        setError(null);
+//     return {
+//         transferOwnership,
+//         isTransferring,
+//         error,
+//     };
+// };
 
-        try {
-            const poolData = await readContract(wagmiConfig, {
-                address: POOL_MANAGER_ADDRESS,
-                abi: PoolManagerABI,
-                functionName: 'pools',
-                args: [poolId] as const,
-            });
+// // Hook for renouncing ownership
+// interface UseRenounceOwnershipReturn {
+//     renounceOwnership: () => Promise<TransactionReceipt>;
+//     isRenouncing: boolean;
+//     error: Error | null;
+// }
 
-            // The returned data is a tuple array, convert it to our Pool type
-            return {
-                maxOrderAmount: poolData[0],
-                lotSize: poolData[1],
-                baseCurrency: poolData[2],
-                quoteCurrency: poolData[3],
-                orderBook: poolData[4],
-            };
-        } catch (err: unknown) {
-            const error = err instanceof Error ? err : new Error('Failed to get pool by ID');
-            setError(error);
-            throw error;
-        } finally {
-            setIsLoading(false);
-        }
-    }, []);
+// export const useRenounceOwnership = (options: BaseOptions = {}): UseRenounceOwnershipReturn => {
+//     const { onSuccess, onError } = options;
+//     const [isRenouncing, setIsRenouncing] = useState(false);
+//     const [error, setError] = useState<Error | null>(null);
 
-    return {
-        getPoolById,
-        isLoading,
-        error,
-    };
-};
+//     const renounceOwnership = useCallback(async (): Promise<TransactionReceipt> => {
+//         setIsRenouncing(true);
+//         setError(null);
 
-// Hook for transferring ownership
-interface TransferOwnershipParams {
-    newOwner: HexAddress;
-}
+//         try {
+//             const hash = await writeContract(wagmiConfig, {
+//                 address: POOL_MANAGER_ADDRESS,
+//                 abi: PoolManagerABI,
+//                 functionName: 'renounceOwnership',
+//             });
 
-interface UseTransferOwnershipReturn {
-    transferOwnership: (params: TransferOwnershipParams) => Promise<TransactionReceipt>;
-    isTransferring: boolean;
-    error: Error | null;
-}
+//             const receipt = await waitForTransaction(wagmiConfig, {
+//                 hash,
+//             });
 
-export const useTransferOwnership = (options: BaseOptions = {}): UseTransferOwnershipReturn => {
-    const { onSuccess, onError } = options;
-    const [isTransferring, setIsTransferring] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
+//             onSuccess?.(receipt);
+//             return receipt;
+//         } catch (err: unknown) {
+//             const error = err instanceof Error ? err : new Error('Failed to renounce ownership');
+//             setError(error);
+//             onError?.(error);
+//             throw error;
+//         } finally {
+//             setIsRenouncing(false);
+//         }
+//     }, [onSuccess, onError]);
 
-    const transferOwnership = useCallback(
-        async ({ newOwner }: TransferOwnershipParams): Promise<TransactionReceipt> => {
-            setIsTransferring(true);
-            setError(null);
-
-            try {
-                const hash = await writeContract(wagmiConfig, {
-                    address: POOL_MANAGER_ADDRESS,
-                    abi: PoolManagerABI,
-                    functionName: 'transferOwnership',
-                    args: [newOwner] as const,
-                });
-
-                const receipt = await waitForTransaction(wagmiConfig, {
-                    hash,
-                });
-
-                onSuccess?.(receipt);
-                return receipt;
-            } catch (err: unknown) {
-                const error = err instanceof Error ? err : new Error('Failed to transfer ownership');
-                setError(error);
-                onError?.(error);
-                throw error;
-            } finally {
-                setIsTransferring(false);
-            }
-        },
-        [onSuccess, onError]
-    );
-
-    return {
-        transferOwnership,
-        isTransferring,
-        error,
-    };
-};
-
-// Hook for renouncing ownership
-interface UseRenounceOwnershipReturn {
-    renounceOwnership: () => Promise<TransactionReceipt>;
-    isRenouncing: boolean;
-    error: Error | null;
-}
-
-export const useRenounceOwnership = (options: BaseOptions = {}): UseRenounceOwnershipReturn => {
-    const { onSuccess, onError } = options;
-    const [isRenouncing, setIsRenouncing] = useState(false);
-    const [error, setError] = useState<Error | null>(null);
-
-    const renounceOwnership = useCallback(async (): Promise<TransactionReceipt> => {
-        setIsRenouncing(true);
-        setError(null);
-
-        try {
-            const hash = await writeContract(wagmiConfig, {
-                address: POOL_MANAGER_ADDRESS,
-                abi: PoolManagerABI,
-                functionName: 'renounceOwnership',
-            });
-
-            const receipt = await waitForTransaction(wagmiConfig, {
-                hash,
-            });
-
-            onSuccess?.(receipt);
-            return receipt;
-        } catch (err: unknown) {
-            const error = err instanceof Error ? err : new Error('Failed to renounce ownership');
-            setError(error);
-            onError?.(error);
-            throw error;
-        } finally {
-            setIsRenouncing(false);
-        }
-    }, [onSuccess, onError]);
-
-    return {
-        renounceOwnership,
-        isRenouncing,
-        error,
-    };
-};
+//     return {
+//         renounceOwnership,
+//         isRenouncing,
+//         error,
+//     };
+// };
